chore: separate validator logic from big blocs
